% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_graph.R
\name{is_osm_graph}
\alias{is_osm_graph}
\title{Test whether an object is an \code{osm_graph}}
\usage{
is_osm_graph(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical scalar.
}
\description{
Test whether an object is an \code{osm_graph}
}
