% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmnxr-package.R
\docType{package}
\name{osmnxr-package}
\alias{osmnxr}
\alias{osmnxr-package}
\title{osmnxr: Download, Model and Analyze 'OpenStreetMap' Street Networks}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

A 'tidyverse'-friendly toolkit, inspired by the 'OSMnx' 'Python' library, to download, model, simplify, analyze and visualize street networks and other geospatial features from 'OpenStreetMap'. Build routable graphs from a place name, address, point or bounding box; simplify topology; compute shortest paths, isochrones and urban metrics (intersection density, circuity, street-orientation entropy, centrality); and export to 'sf', 'sfnetworks' and 'MapLibre'. Heavy graph computation is performed by a bundled 'Rust' core.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StrategicProjects/osmnxr}
  \item \url{https://strategicprojects.github.io/osmnxr/}
  \item Report bugs at \url{https://github.com/StrategicProjects/osmnxr/issues}
}

}
\author{
\strong{Maintainer}: Andre Leite \email{leite@castlab.org}

Authors:
\itemize{
  \item Andre Leite \email{leite@castlab.org}
  \item Marcos Wasiliew \email{marcos.wasiliew@gmail.com} (\href{https://orcid.org/0009-0004-4694-3159}{ORCID})
  \item Hugo Vasconcelos \email{hugo.vasconcelos@ufpe.br}
  \item Carlos Amorim \email{carlos.agaf@ufpe.br}
  \item Diogo Bezerra \email{diogo.bezerra@ufpe.br}
  \item Júlia Nascimento Barreto \email{juliabarreto@gd.seplag.pe.gov.br} (\href{https://orcid.org/0009-0004-2851-7770}{ORCID})
}

Other contributors:
\itemize{
  \item StrategicProjects [copyright holder, funder]
  \item The extendr authors (Bundled Rust crates extendr-api, extendr-ffi, extendr-macros) [copyright holder]
  \item David Tolnay (Bundled Rust crates proc-macro2, quote, syn, paste, readonly, unicode-ident) [copyright holder]
  \item Alex Crichton (Bundled Rust crate proc-macro2) [copyright holder]
  \item Marvin Loebel (Bundled Rust crate lazy_static) [copyright holder]
  \item Aleksey Kladov (Bundled Rust crate once_cell) [copyright holder]
  \item Unicode, Inc. (Bundled Rust crate unicode-ident (Unicode-3.0 data tables)) [copyright holder]
}

}
\keyword{internal}
