% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop.R
\name{ox_as_dodgr}
\alias{ox_as_dodgr}
\title{Convert to a \code{dodgr} graph}
\usage{
ox_as_dodgr(g, weight = "length")
}
\arguments{
\item{g}{An \link[=new_osm_graph]{osm_graph}.}

\item{weight}{Edge column used as the distance/weight \code{d}. Default
\code{"length"}.}
}
\value{
A \code{data.frame} \code{dodgr} graph.
}
\description{
Returns a \code{data.frame} in the column layout expected by the \code{dodgr} routing
package (\code{from_id}, \code{from_lon}, \code{from_lat}, \code{to_id}, \code{to_lon}, \code{to_lat},
\code{d}), suitable for \code{dodgr::dodgr_dists()} and friends.
}
\examples{
\dontshow{if (rlang::is_installed("dodgr")) withAutoprint(\{ # examplesIf}
g <- example_osm_graph()
head(ox_as_dodgr(g))
\dontshow{\}) # examplesIf}
}
