% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_graph.R
\name{ox_as_sf}
\alias{ox_as_sf}
\title{Extract \code{sf} nodes and edges from an \code{osm_graph}}
\usage{
ox_as_sf(g)
}
\arguments{
\item{g}{An \code{osm_graph}.}
}
\value{
A named list with \code{sf} elements \code{nodes} and \code{edges}.
}
\description{
Extract \code{sf} nodes and edges from an \code{osm_graph}
}
\examples{
g <- example_osm_graph()
parts <- ox_as_sf(g)
parts$edges
}
