% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interop.R
\name{ox_as_sfnetwork}
\alias{ox_as_sfnetwork}
\title{Convert to an \code{sfnetwork}}
\usage{
ox_as_sfnetwork(g, directed = TRUE)
}
\arguments{
\item{g}{An \link[=new_osm_graph]{osm_graph}.}

\item{directed}{Build a directed network. Default \code{TRUE}.}
}
\value{
An \code{sfnetwork}.
}
\description{
Returns the graph as a \code{sfnetworks::sfnetwork()} object, ready for the
\code{sfnetworks}/\code{tidygraph} spatial-network workflow.
}
\examples{
\dontshow{if (rlang::is_installed("sfnetworks")) withAutoprint(\{ # examplesIf}
g <- example_osm_graph()
ox_as_sfnetwork(g)
\dontshow{\}) # examplesIf}
}
