% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{ox_clear_cache}
\alias{ox_clear_cache}
\title{Clear the session cache}
\usage{
ox_clear_cache()
}
\value{
Invisibly \code{NULL}.
}
\description{
Empties the in-memory cache of downloaded OpenStreetMap responses.
}
\examples{
ox_clear_cache()
}
