% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ox_graph_from_address}
\alias{ox_graph_from_address}
\title{Download a street network around an address}
\usage{
ox_graph_from_address(address, dist = 1000, network_type = "drive")
}
\arguments{
\item{address}{A street address.}

\item{dist}{Buffer half-width in metres. Default \code{1000}.}

\item{network_type}{One of \code{"drive"}, \code{"walk"}, \code{"bike"} or \code{"all"}.}
}
\value{
An \link[=new_osm_graph]{osm_graph}.
}
\description{
Download a street network around an address
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
g <- ox_graph_from_address("Marco Zero, Recife", dist = 600)
\dontshow{\}) # examplesIf}
}
