% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ox_graph_from_bbox}
\alias{ox_graph_from_bbox}
\title{Download a street network within a bounding box}
\usage{
ox_graph_from_bbox(bbox, network_type = "drive")
}
\arguments{
\item{bbox}{Numeric vector \code{c(xmin, ymin, xmax, ymax)} in longitude/latitude
(EPSG:4326).}

\item{network_type}{One of \code{"drive"}, \code{"walk"}, \code{"bike"} or \code{"all"}.}
}
\value{
An \link[=new_osm_graph]{osm_graph}.
}
\description{
Download a street network within a bounding box
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
bbox <- c(-34.91, -8.07, -34.87, -8.04)
g <- ox_graph_from_bbox(bbox, network_type = "drive")
\dontshow{\}) # examplesIf}
}
