% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ox_orientation_entropy}
\alias{ox_orientation_entropy}
\title{Street-orientation entropy}
\usage{
ox_orientation_entropy(g, num_bins = 36)
}
\arguments{
\item{g}{An \link[=new_osm_graph]{osm_graph}.}

\item{num_bins}{Number of equal bearing sectors over \verb{[0, 360)}. Default \code{36}.}
}
\value{
A numeric scalar (entropy in nats).
}
\description{
Shannon entropy (in nats) of the distribution of edge compass bearings,
binned into equal sectors. Higher values indicate a more disordered
(organic) network; lower values a more ordered (gridiron) one.
}
\examples{
g <- example_osm_graph()
ox_orientation_entropy(g)
}
