% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{ox_shortest_path}
\alias{ox_shortest_path}
\title{Shortest path between two nodes}
\usage{
ox_shortest_path(g, from, to, weight = "length")
}
\arguments{
\item{g}{An \link[=new_osm_graph]{osm_graph}.}

\item{from, to}{Node \code{osmid}s (as returned by \code{\link[=ox_nearest_nodes]{ox_nearest_nodes()}}).}

\item{weight}{Edge column used as weight. Default \code{"length"}.}
}
\value{
A vector of node \code{osmid}s describing the path (length 0 if the
target is unreachable).
}
\description{
Computes the minimum-weight path from \code{from} to \code{to} using Dijkstra's
algorithm in the Rust core.
}
\examples{
g <- example_osm_graph()
from <- ox_nearest_nodes(g, 0, 0)
to <- ox_nearest_nodes(g, 300, 300)
ox_shortest_path(g, from, to)
}
