% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_basic_stats}
\alias{rs_basic_stats}
\title{Basic network statistics from edge endpoints (0-based) and weights.}
\usage{
rs_basic_stats(from, to, weight, n_nodes)
}
\description{
Basic network statistics from edge endpoints (0-based) and weights.
}
\keyword{internal}
