% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_bearings}
\alias{rs_bearings}
\title{Initial compass bearings (degrees) for parallel coordinate vectors.}
\usage{
rs_bearings(lat1, lon1, lat2, lon2)
}
\description{
Initial compass bearings (degrees) for parallel coordinate vectors.
}
\keyword{internal}
