% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_closeness}
\alias{rs_closeness}
\title{Closeness centrality for every node.}
\usage{
rs_closeness(from, to, weight, n_nodes, normalized)
}
\description{
Closeness centrality for every node.
}
\keyword{internal}
