% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_connected_components}
\alias{rs_connected_components}
\title{Connected-component label (0-based root index) for each node, given
undirected \code{a}--\code{b} adjacency pairs.}
\usage{
rs_connected_components(a, b, n_nodes)
}
\description{
Connected-component label (0-based root index) for each node, given
undirected \code{a}--\code{b} adjacency pairs.
}
\keyword{internal}
