% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_dijkstra}
\alias{rs_dijkstra}
\title{Single-source shortest distances (length \code{n_nodes}); \code{Inf} if unreachable.}
\usage{
rs_dijkstra(from, to, weight, n_nodes, source)
}
\description{
Single-source shortest distances (length \code{n_nodes}); \code{Inf} if unreachable.
}
\keyword{internal}
