% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_distance_matrix}
\alias{rs_distance_matrix}
\title{Row-major shortest-distance matrix from each source to each target
(0-based indices). Length is \code{sources * targets}.}
\usage{
rs_distance_matrix(from, to, weight, n_nodes, sources, targets)
}
\description{
Row-major shortest-distance matrix from each source to each target
(0-based indices). Length is \code{sources * targets}.
}
\keyword{internal}
