% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_k_shortest_paths}
\alias{rs_k_shortest_paths}
\title{Yen's k loopless shortest paths. Returns a list with \code{paths} (a list of
0-based node-index vectors) and \code{costs} (numeric).}
\usage{
rs_k_shortest_paths(from, to, weight, n_nodes, source, target, k)
}
\description{
Yen's k loopless shortest paths. Returns a list with \code{paths} (a list of
0-based node-index vectors) and \code{costs} (numeric).
}
\keyword{internal}
