% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_orientation_entropy}
\alias{rs_orientation_entropy}
\title{Shannon entropy of binned edge bearings (degrees).}
\usage{
rs_orientation_entropy(bearings, num_bins)
}
\description{
Shannon entropy of binned edge bearings (degrees).
}
\keyword{internal}
