% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rs_shortest_path}
\alias{rs_shortest_path}
\title{Shortest path (0-based node indices) between \code{source} and \code{target}.
Returns an empty vector when the target is unreachable.}
\usage{
rs_shortest_path(from, to, weight, n_nodes, source, target)
}
\description{
Shortest path (0-based node indices) between \code{source} and \code{target}.
Returns an empty vector when the target is unreachable.
}
\keyword{internal}
